IOS-5135: Fix naming ambiguity
