+++ Tangem/Modules/Main/MainBottomSheet/Header/MainBottomSheetHeaderView.swift
@@ -9,35 +9,9 @@
 import SwiftUI
 
 struct MainBottomSheetHeaderView: View {
-    @Binding private var searchText: String
-    private let textFieldAllowsHitTesting: Bool
-
-    init(
-        searchText: Binding<String>,
-        textFieldAllowsHitTesting: Bool
-    ) {
-        _searchText = searchText
-        self.textFieldAllowsHitTesting = textFieldAllowsHitTesting
-    }
+    @ObservedObject var viewModel: MainBottomSheetHeaderViewModel
 
     var body: some View {
-        TextField(Localization.commonSearch, text: $searchText)
-            .allowsHitTesting(textFieldAllowsHitTesting)
-            .frame(height: 46.0)
-            .padding(.horizontal, 12.0)
-            .background(Colors.Field.primary)
-            .cornerRadius(14.0)
-            .padding(.horizontal, 16.0)
-            .padding(.top, Constants.verticalInset)
-            .padding(.bottom, max(UIApplication.safeAreaInsets.bottom, Constants.verticalInset))
-            .background(Colors.Background.primary)
-    }
-}
-
-// MARK: - Constants
-
-private extension MainBottomSheetHeaderView {
-    enum Constants {
-        static let verticalInset = 20.0
+        MainBottomSheetHeaderInputView(searchText: $viewModel.enteredSearchText, textFieldAllowsHitTesting: true)
     }
 }
