+++ Tangem/Modules/Main/MainBottomSheet/Navigation/MainBottomSheetHeaderCoordinatorView.swift
@@ -15,7 +15,7 @@ struct MainBottomSheetHeaderCoordinatorView: CoordinatorView {
 
     var body: some View {
         if let viewModel = coordinator.headerViewModel {
-            MainBottomSheetHeaderContainerView(viewModel: viewModel)
+            MainBottomSheetHeaderView(viewModel: viewModel)
         }
     }
 }
